requirements.txt
Flask==2.2.5
Flask-CORS==4.0.0
Pillow==9.5.0
qrcode==7.3.1
gunicorn==20.1.0
